refactor(gateways): move the transport logic out of the constants file
The file had grown to five constants and eight functions. The existing
gateway-v2-fns imports the service, which imports the constants, so moving
them there would have closed a cycle; gateway-v2-transport-fns depends on
nothing but the constants.

The enum stays with the constants, since the routers and audit-log types
need it and have no business reaching into a transport module. The
TGatewayTransport alias went with the move: it aliased the enum exactly
and nothing outside the moved code used it.
