+++ backend/src/ee/services/gateway-pool/gateway-pool-membership-dal.ts
@@ -3,7 +3,7 @@ import { TableName, TGatewaysV2 } from "@app/db/schemas";
 import { DatabaseError } from "@app/lib/errors";
 import { ormify } from "@app/lib/knex";
 
-import { buildGatewayReachableSql } from "../gateway-v2/gateway-v2-constants";
+import { buildGatewayReachableSql } from "../gateway-v2/gateway-v2-transport-fns";
 
 export type TGatewayPoolMembershipDALFactory = ReturnType<typeof gatewayPoolMembershipDalFactory>;
